chore: move dapa client to stage in-out coz it is only used in that module
