+++ cumulus_lambda_functions/stage_in_out/search_granules_unity.py
@@ -2,8 +2,7 @@ import json
 import logging
 import os
 
-from cumulus_lambda_functions.cumulus_dapa_client.dapa_client import DapaClient
-from cumulus_lambda_functions.cumulus_stac.stac_utils import StacUtils
+from cumulus_lambda_functions.stage_in_out.dapa_client import DapaClient
 from cumulus_lambda_functions.stage_in_out.search_granules_abstract import SearchGranulesAbstract
 
 LOGGER = logging.getLogger(__name__)
